requirements.txt
openai
streamlit
pandas
TTS
torch
python-dotenv
openpyxl